Workaround for the fact that the HeapNumber value is not aligned to 8-byte boundary.
MIPS and gcc use ldc1 and sdc1 instructions to load/store doubles from memory to FPU 
registers. These require the targets to be 64-aligned. This is not the case in v8 today.

Our codegen uses a pair of 32-bit load/store instructions, so that we do not care
about alignment. But we cannot control how gcc compiles c++ code. 

This workaround changes the HeapNumber accessors so memory is accessed with 32-bit
instructions.

It is a hack, and long-term we should fix the alignment of HeapNumbers, and revert
this change.