Change the Label end-of-chain sentinel to fix earley-boyer.js.
Prior to this change, no function could be larger than 128K,
due to implementation of the end-of-chain mechanism, which
recorded an offset to start of function. (Relative branches
in Mips have 16 bit offset, representing +/- 128K bytes).

With this change, a label and all its branches must be within
+/- 128K range, but total function size is not limited.

This change does not correctly handle reg-exp use of
label_at_put().