Calling builtins via CEntry stub _must_ use jalr t9.
The ABI mandates use of t9 register for C calling convention in
position-independent-code. This allows correct adjustment of 'gp'
register to point to the GOT for this compilation/link unit.

This is not important inside our own generated code, but is critical
at the interface to gcc-generated code, as called thru this stub.