Fix for the compacting garbage collection bug on mips.
The problem was incorrect updating of embedded object pointers in code,
when code objects were relocated. Specifically, the
UpdatingVisitor::VisitPointer() in mark-compact.cc was updating objects
with indirect pointer.

Had to backport some changes from google code relating to object visitors,
and then refactor things a bit to allow this Visitor to use RelocInfo,
and update the pointers with RelocInfo::set_target_object().

This has been tested on ia32 and arm, and is fine.