Fix printing of external-reference calls in code disassembles and reloc-info.
The loading of a pointer to external-reference was mistakenly excluded from
the disassembler listing. Also a bug prevented the name of the routine from being
looked up. Now we get the routine names listed inline on the disassembly like this:

0xd3b3d0   496  3c0500a0       lui    a1, 0xa0               ;; external reference (Runtime::ToBool)
0xd3b3d4   500  34a53bc4       ori    a1, a1, 0x3bc4
0xd3b3d8   504  3c1900c2       lui    t9, 0xc2               ;; code: STUB, CEntry, minor: 0
0xd3b3dc   508  37390120       ori    t9, t9, 0x120
0xd3b3e0   512  0320f809       jalr t9

(There was no advantage to excluding ExternalReference calls from MustUseReg() routine,
as they can never be encoded in 16 bits regardless of use of Serializer. All it did
was supress disassembly.)